requirements.txt
openai>=1.0.0
langchain>=0.1.14
langchain-openai>=0.0.8
langchain-text-splitters>=0.0.1
langchain-core>=0.1.30
langchain-neo4j>=0.1.3
neo4j>=5.13.0
requests>=2.31.0
numpy>=1.24.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv
streamlit>=1.36.0
